refactor proc package
The proc package contained a lot of duplicated code.

This commit extracts that code into some more generic methods.
