+++ pkg/proc/types.go
@@ -15,58 +15,85 @@ const (
 
 type Runner struct {
 	IgnitionConfig *config.Ignition
-	jobs           []*Job
-	bootJobs       []*BootJob
+
+	jobs     []*Job
+	lazyJobs []*LazyJob
+	bootJobs []*BootJob
 }
 
-type BootJob struct {
-	Config  *config.BootJobConfig
+type BaseJob struct {
+	Config *config.BaseJobConfig
+
 	cmd     *exec.Cmd
-	process *os.Process
-	timeout time.Duration
 }
 
 type Job struct {
-	Config        *config.JobConfig
-	watchingFiles map[string]time.Time
-	cmd           *exec.Cmd
-	process       *os.Process
+	BaseJob
+	Config *config.JobConfig
 
-	lazyStartLock sync.Mutex
+	watchingFiles map[string]time.Time
+}
+
+type BootJob struct {
+	BaseJob
+	Config *config.BootJobConfig
+
+	timeout time.Duration
+}
+
+type LazyJob struct {
+	BaseJob
+	Config *config.JobConfig
+
+	process *os.Process
+
+	lazyStartLock     sync.Mutex
+	activeConnections uint32
 
 	spinUpTimeout        time.Duration
 	coolDownTimeout      time.Duration
 	lastConnectionClosed time.Time
-	activeConnections    uint32
 }
 
-func NewJob(c *config.JobConfig) (*Job, error) {
+func NewJob(c *config.JobConfig) *Job {
 	j := Job{
+		BaseJob: BaseJob{
+			Config: &c.BaseJobConfig,
+		},
 		Config: c,
 	}
 
-	if c.Laziness != nil {
-		if c.Laziness.SpinUpTimeout != "" {
-			t, err := time.ParseDuration(c.Laziness.SpinUpTimeout)
-			if err != nil {
-				return nil, err
-			}
+	return &j
+}
 
-			j.spinUpTimeout = t
-		} else {
-			j.spinUpTimeout = 5 * time.Second
+func NewLazyJob(c *config.JobConfig) (*LazyJob, error) {
+	j := LazyJob{
+		BaseJob: BaseJob{
+			Config: &c.BaseJobConfig,
+		},
+		Config: c,
+	}
+
+	if c.Laziness.SpinUpTimeout != "" {
+		t, err := time.ParseDuration(c.Laziness.SpinUpTimeout)
+		if err != nil {
+			return nil, err
 		}
 
-		if c.Laziness.CoolDownTimeout != "" {
-			t, err := time.ParseDuration(c.Laziness.CoolDownTimeout)
-			if err != nil {
-				return nil, err
-			}
+		j.spinUpTimeout = t
+	} else {
+		j.spinUpTimeout = 5 * time.Second
+	}
 
-			j.coolDownTimeout = t
-		} else {
-			j.coolDownTimeout = 15 * time.Minute
+	if c.Laziness.CoolDownTimeout != "" {
+		t, err := time.ParseDuration(c.Laziness.CoolDownTimeout)
+		if err != nil {
+			return nil, err
 		}
+
+		j.coolDownTimeout = t
+	} else {
+		j.coolDownTimeout = 15 * time.Minute
 	}
 
 	return &j, nil
@@ -74,6 +101,9 @@ func NewJob(c *config.JobConfig) (*Job, error) {
 
 func NewBootJob(c *config.BootJobConfig) (*BootJob, error) {
 	bj := BootJob{
+		BaseJob: BaseJob{
+			Config: &c.BaseJobConfig,
+		},
 		Config: c,
 	}
